MySQL Initial Implementation: Bootstrap Checkin
This is the beginnings of the MySQL project. Ignore everything in the providers directory; it's just auto-generated code right now.

This includes a .MOF file (reviewed by PM), the provider registration file, a configure program, and a Makefile. Note that a lot of the targets in Makefile are untested, although "make all" does work to build the provider library, along with the OMI targets to build the providers and the registration file. Specifically, tests and kits are not yet working, I'll get those going when I get something ready.